Enhanced robustness of database code
Added a comment to line 28 indicating purpose of initialSetup.

Added a check on line 39 to see if the database connect in still
current. It may not be when someone is playing with the code
interactively. This is not an elegant fix, but is easily understood.

Added a check on line 58 to make sure readHTMLTable() returned some
data.